Refactor component structure
